domains refactored
