+++ src/main/java/com/epam/lab/library/Application.java
@@ -1,8 +1,5 @@
 package com.epam.lab.library;
 
-import com.epam.lab.library.DAO.Implementations.BookDAOImpl;
-import com.epam.lab.library.domain.Book;
-
 public class Application {
 
     public static void main(String[] args) {
