+++ src/main/java/com/epam/lab/library/domain/Role.java
@@ -4,15 +4,14 @@ import java.util.Set;
 
 public class Role {
 
-    private int id;
+    private Integer id;
     private String name;
-    private Set<User> users;
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public Role setId(int id) {
+    public Role setId(Integer id) {
         this.id = id;
         return this;
     }
@@ -23,15 +22,6 @@ public class Role {
 
     public Role setName(String name) {
         this.name = name;
-        return this;
-    }
-
-    public Set<User> getUsers() {
-        return users;
-    }
-
-    public Role setUsers(Set<User> users) {
-        this.users = users;
         return this;
     }
 }
