+++ src/main/java/com/epam/lab/library/domain/Status.java
@@ -4,15 +4,14 @@ import java.util.Set;
 
 public class Status {
 
-    private int id;
+    private Integer id;
     private String name;
-    private Set<Item> items;
 
-    public int getId() {
+    public Integer getId() {
         return id;
     }
 
-    public Status setId(int id) {
+    public Status setId(Integer id) {
         this.id = id;
         return this;
     }
@@ -23,15 +22,6 @@ public class Status {
 
     public Status setName(String name) {
         this.name = name;
-        return this;
-    }
-
-    public Set<Item> getItems() {
-        return items;
-    }
-
-    public Status setItems(Set<Item> items) {
-        this.items = items;
         return this;
     }
 }
